added continous scale sentiment values per tweet
chunk of code was added for continuous sentistrength scaling. end result adds positive and negative column to made data frame variable (d). To use in later analysis, the columns can be added to the select() function in the creation of sentiment_states -- they can then be worked with in sentiment_means